+++ src/core/requestOptions.js
@@ -1,6 +1,6 @@
+// Can't import json like require
 // import { version } from '../package.json';
 const version = 'test';
-import { getConfig } from './config';
 
 const DEFAULT_USER_AGENT_HEADERS = {
   'User-Agent': `HubSpot CLI/${version}`,
@@ -10,8 +10,7 @@ const DEFAULT_USER_AGENT_HEADERS = {
 
 const getRequestOptions = (options = {}, requestOptions = {}) => {
   const { env } = options;
-  const { httpTimeout, httpUseLocalhost } = getConfig();
-  console.log('r2', requestOptions);
+
   return {
     headers: {
       ...DEFAULT_USER_AGENT_HEADERS,
